Refactor LigneAchat and Transaction services for modularity
Split `TransactionAchatAndLinesService` into `LigneAchatService` for better separation of concerns. Updated mappers for streamlined and reusable entity-to-DTO conversions.

+++ src/main/java/com/projet/equipement/dto/ligneVente/LigneVenteUpdateDto.java
@@ -13,11 +13,11 @@ public class LigneVenteUpdateDto {
 
     private Integer quantite;
 
-    private Integer venteId;
+    private Long venteId;
 
     private Boolean actif;
 
-    private Integer produitId;
+    private Long produitId;
 
     private Long formatVenteId;
